move stuff around, add test for locked files in search response
